% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.r
\name{cluster}
\alias{cluster}
\title{Data-clustering}
\usage{
cluster(data, ...)
}
\arguments{
\item{df}{Dataframe to be clustered. Method also possible with vectors.}

\item{group}{Defaults to \code{c()}. Specificies columns, by which data is to be preliminarily divided into groups, within which the clusters are to be built.}

\item{by}{Specifies the column(s) for geometric data, according to which the clusters are to be built.}

\item{isnear}{A function. This function operates pairs of entries in the columns with geometric data and returns \code{TRUE}/\code{FALSE} if entries are near. Defaults to a Euclidian metric.}

\item{dist}{Defaults to \code{Inf}. If the default euclidean metric is used for \code{isnear}, this is the maximum tolerated distance between geometric data.}

\item{strict}{Defaults to \code{TRUE}. If the default euclidean metric is used for \code{isnear}, this sets the proximity to be a strict \code{< dist} or else \code{<= dist}.}

\item{clustername}{Defaults to \code{'cluster'}. Running \code{df \%>\% clusterby(...)} returns a data frame, which extends \code{df} by 1 column with this name. This column tags the clusters by a unique index.}

\item{min_cluster_size}{Defaults to \code{0}. If a cluster is smaller than this, it will not be viewed as a cluster.}

\item{max_cluster_size}{Defaults to \code{Inf}. If a cluster is larger than this, it will be broken up into smaller pieces.}

\item{split}{Defaults to \code{FALSE}. If set to \code{TRUE}, then the output will be group the tibble data by cluster (equivalent to performing \code{\%>\% group_by(...)}).}

\item{tib}{Tibble data which has been grouped.}
}
\description{
This package contains methods, which enables clustering in dataframes. Particularly useful for bio-mathematics, cognitive sciences, etc.
}
\details{
\code{cluster(df, ...)}
}
\examples{
gene \%>\% cluster(by='position', group=c('gene','actve'), dist=400, strict=FALSE, clustername='tag');
protein3d \%>\% cluster(by=c('x','y','z'), group=c('gene','actve'), dist=2.5, max=5000);
\\code{summarise(tib, ...)}
tib \%>\% summarise(concentration=mean, names=c('set',';'), attributes=c('list',';'));
}
\keyword{cluster}
\keyword{clustering}
\keyword{gene}
