% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.r
\name{clusterby}
\alias{clusterby}
\title{Data-clustering}
\usage{
clusterby(data, ...)
}
\arguments{
\item{df}{Dataframe to be clustered. Method also possible with vectors.}

\item{by}{Specifies the column name for geometric data, according to which the clusters are to be built.}

\item{group}{Defaults to \code{c()}. Specificies columns, by which data is to be preliminarily divided into groups, within which the clusters are to be built.}

\item{dist}{Defaults to \code{Inf}. For a column \code{x} of positions, a cluster is a subset \code{y} which satisfies \bold{(i)} between every two points there is a path within \code{y} whose steps are 'small'; \bold{(ii)} y is maximal under condition (i). 'Small' here means the the distance between two points is strictly smaller than \code{d}.}

\item{strict}{Defaults to \code{TRUE}. If set to \code{FALSE} the 'smaller than' is replaced by 'smaller than or equal to'.}

\item{clustername}{Defaults to 'cluster'. Running \code{df \%>\% clusterby(...)} returns a data frame, which extends \code{df} by 1 column with this name. This column tags the clusters by a unique index.}

\item{dim}{New feature to be added. Allow points to be arbitrary data, allow input of a distance function/neighbourhoodscheme.}
}
\description{
This package contains methods, which enables clustering in dataframes. Particularly useful for bio-mathematics, cognitive sciences, etc.
}
\details{
\code{clusterby(df, ...)}
}
\examples{
df \%>\% clusterby(by='position', group=c('gene','actve'), dist=400, strict=FALSE, clustername='tag');
p \%>\% clusterby(dist=10); # p is a numeric vector
}
\keyword{cluster}
\keyword{clustering}
\keyword{gene}
