% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.r
\name{summarise}
\alias{summarise}
\title{Data-clustering | summarise}
\usage{
summarise(tib, ...)
}
\arguments{
\item{tib}{tibble data which has been grouped.}
}
\description{
Provides customisable summary method tailored for clustering.
}
\details{
\code{summarise(tib, ...)}
}
\examples{
tib \%>\% summarise(concentration=mean, names=c('set',';'), attributes=c('list',';'));
}
\keyword{cluster}
\keyword{summarise}
