% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.r
\name{summarise}
\alias{summarise}
\title{\code{summarise(tib, ...)}}
\usage{
summarise(tib, ...)
}
\arguments{
\item{tib}{Tibble data which has been grouped.}
}
\description{
\code{summarise(tib, ...)}
}
\examples{
tib \%>\% summarise(concentration=mean, names=c('set',';'), attributes=c('list',';'));
}
\keyword{cluster}
\keyword{clustering}
\keyword{gene}
